requirements.txt
pycryptodome
requests